FIXED: a few warnings, which made up over 70% of all warning-scroll using gcc/g++; NEL-25
git-svn-id: svn://svn.code.sf.net/p/nel/code/trunk@254 bf685bcc-7839-4fcd-9254-36823579bede

+++ nel/src/3d/index_buffer.h
@@ -435,7 +435,7 @@ private:
 	inline void	unlock () const;
 
 	// Reset the touch flags
-	void		resetTouchFlags() {_InternalFlags &= ~TouchedAll;}
+	void		resetTouchFlags() {_InternalFlags &= (uint16)(~TouchedAll);}
 
 	// Force non resident memory
 	void		restaureNonResidentMemory();
